fix/sizeM2-format-crash

+++ app/src/main/java/com/example/g46_kotlin/features/map/data/remote/dto/PropertyDto.kt
@@ -31,7 +31,7 @@ data class PropertyDto(
     val city: String,
     val latitude: Double,
     val longitude: Double,
-    val sizeM2: Int,
+    val sizeM2: Double,
     val bedrooms: Int,
     val bathrooms: Int,
     val furnished: Boolean,
